// src/app/models/page-content.ts
export interface PageContent {
    siteName: string;
    homeMainTitle: string;
    homeMainText: string;
    homeTagLine: string;
    homeSub1Title: string;
    homeSub1Text: string;
    homeSub2Title: string;
    homeSub2Text: string;
    homeSub3Title: string;
    homeSub3Text: string;
    contactText: string;
    contactEmail: string;
    contactPhone: string;
}